first pass

+++ apps/platform/pkg/adapt/postgresio/save.go
@@ -10,10 +10,10 @@ import (
 	"github.com/thecloudmasters/uesio/pkg/sess"
 )
 
-const INSERT_QUERY = "INSERT INTO public.data (id,uniquekey,collection,tenant,autonumber,fields) VALUES ($1,$2,$3,$4,$5,$6)"
-const UPDATE_QUERY = "UPDATE public.data SET uniquekey = $2, fields = fields || $4 WHERE id = $1 and collection = $3"
-const DELETE_QUERY = "DELETE FROM public.data WHERE id = ANY($1) and collection = $2"
-const TOKEN_DELETE_QUERY = "DELETE FROM public.tokens WHERE recordid = ANY($1)"
+const INSERT_QUERY = "INSERT INTO public.data (id,uniquekey,owner,collection,tenant,autonumber,fields) VALUES ($1,$2,$3,$4,$5,$6,$7)"
+const UPDATE_QUERY = "UPDATE public.data SET uniquekey = $2, owner = $3, fields = fields || $6 WHERE id = $1 and collection = $4 and tenant = $5"
+const DELETE_QUERY = "DELETE FROM public.data WHERE id = ANY($1) and collection = $2 and tenant = $3"
+const TOKEN_DELETE_QUERY = "DELETE FROM public.tokens WHERE recordid = ANY($1) and collection = $2 and tenant = $3"
 const TOKEN_INSERT_QUERY = "INSERT INTO public.tokens (recordid,token,collection,tenant,readonly) VALUES ($1,$2,$3,$4,$5)"
 
 func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
@@ -36,13 +36,18 @@ func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 		if err != nil {
 			return err
 		}
+
+		ownerID, err := change.GetOwnerID()
+		if err != nil {
+			return err
+		}
 		fullRecordID := change.IDValue
 		uniqueID := change.UniqueKey
 
 		if change.IsNew {
-			batch.Queue(INSERT_QUERY, fullRecordID, uniqueID, collectionName, tenantID, change.Autonumber, fieldJSON)
+			batch.Queue(INSERT_QUERY, fullRecordID, uniqueID, ownerID, collectionName, tenantID, change.Autonumber, fieldJSON)
 		} else {
-			batch.Queue(UPDATE_QUERY, fullRecordID, uniqueID, collectionName, fieldJSON)
+			batch.Queue(UPDATE_QUERY, fullRecordID, uniqueID, ownerID, collectionName, tenantID, fieldJSON)
 		}
 
 		if request.Metadata.IsWriteProtected() {
@@ -64,11 +69,11 @@ func (c *Connection) Save(request *adapt.SaveOp, session *sess.Session) error {
 		for i, delete := range request.Deletes {
 			deleteIDs[i] = delete.IDValue
 		}
-		batch.Queue(DELETE_QUERY, deleteIDs, collectionName)
+		batch.Queue(DELETE_QUERY, deleteIDs, collectionName, tenantID)
 	}
 
 	if len(resetTokenIDs) > 0 {
-		batch.Queue(TOKEN_DELETE_QUERY, resetTokenIDs)
+		batch.Queue(TOKEN_DELETE_QUERY, resetTokenIDs, collectionName, tenantID)
 	}
 
 	if len(readWriteTokens) > 0 {
